+++ lib/happypass.dart
@@ -3,3 +3,4 @@ library happypass;
 export 'src/http.dart';
 export 'src/form_data.dart';
 export 'src/multi_part.dart';
+export 'src/http_mock_interceptor.dart';